requirements.txt
torch
torchvision
pillow
pandas
scikit-learn
seaborn
matplotlib
glob2
os-win
numpy
tqdm
aiogram
python-dotenv